+++ docs/release-notes.md
@@ -25,13 +25,14 @@ See Firefly docker guidelines [here](firefly-docker.md).
   - Choice of table with VO Tables
   - PDF and TAR are recognized and downloadable.
   
-  
+##### _API_
+- Add fixed column feature to table API ([Firefly-442](https://github.com/Caltech-IPAC/firefly/pull/941)))
+    
 ##### _Pull Request in this release_
   
   - [bug fixes](https://github.com/Caltech-IPAC/firefly/pulls?q=is%3Aclosed+is%3Apr+label%3Abug+milestone%3A2020.1)
   - [All PRs](https://github.com/Caltech-IPAC/firefly/pulls?utf8=%E2%9C%93&q=is%3Aclosed+is%3Apr+)
                         
-
 ### Version 2019.4
 
 ##### _Docker tags, releases and patches_